+++ src/Http/UploadedFile.php
@@ -144,71 +144,83 @@ class UploadedFile implements UploadedFileInterface
     }
 
     /**
-     * Parses the $_FILES into multiple \File instances.
+     * Returns a \UploadedFile instance from $_FILES.
      *
-     * @param  array $uploaded
+     * @param  array $file
+     * @return \Psr\Http\Message\UploadedFileInterface
+     */
+    public static function create(array $file)
+    {
+        if (is_array($file['tmp_name']) === false) {
+            $tmp = $file['tmp_name'];
+
+            $size = $file['size'];
+
+            $error = $file['error'];
+
+            $original = $file['name'];
+
+            $type = $file['type'];
+
+            return new UploadedFile($tmp, $size, $error, $original, $type);
+        }
+
+        return self::nested($file);
+    }
+
+    /**
+     * Returns an array of \UploadedFile instances from $_FILES.
+     *
+     * @param  array $file
+     * @return \Psr\Http\Message\UploadedFileInterface[]
+     */
+    public static function nested(array $files)
+    {
+        $normalized = array();
+
+        foreach (array_keys($files['tmp_name']) as $key) {
+            $file = array('tmp_name' => $files['tmp_name'][$key]);
+
+            $file['size'] = $files['size'][$key];
+
+            $file['error'] = $files['error'][$key];
+
+            $file['name'] = $files['name'][$key];
+
+            $file['type'] = $files['type'][$key];
+
+            $normalized[$key] = self::create($file);
+        }
+
+        return $normalized;
+    }
+
+    /**
+     * Parses the $_FILES into multiple \UploadedFile instances.
+     *
      * @param  array $files
-     * @return \Zapheus\Http\Message\FileInterface[]
+     * @return \Psr\Http\Message\UploadedFileInterface[]
      */
-    public static function normalize(array $uploaded, $files = array())
+    public static function normalize(array $files)
     {
-        foreach (self::diverse($uploaded) as $name => $file) {
-            list($files[$name], $items) = array($file, array());
+        $normalized = array();
 
-            if (isset($file['name']) === true) {
-                foreach ($file['name'] as $key => $value) {
-                    $instance = self::create($file, $key);
+        foreach ((array) $files as $key => $value) {
+            $instanceof = $value instanceof UploadedFileInterface;
 
-                    $items[] = $instance;
-                }
+            if (isset($value['tmp_name'])) {
+                $file = self::create($value);
 
-                $files[$name] = (array) $items;
+                $normalized[$key] = $file;
+            } elseif (is_array($value) === true) {
+                $items = self::normalize($value);
+
+                $normalized[$key] = (array) $items;
             }
+
+            $instanceof === true && $normalized[$key] = $value;
         }
 
-        return $files;
-    }
-
-    /**
-     * Creates a new UploadedFile instance.
-     *
-     * @param  array   $file
-     * @param  integer $key
-     * @return \Psr\Http\Message\UploadedFile
-     */
-    protected static function create(array $file, $key)
-    {
-        $tmp = $file['tmp_name'][$key];
-
-        $size = $file['size'][$key];
-
-        $error = $file['error'][$key];
-
-        $original = $file['name'][$key];
-
-        $type = $file['type'][$key];
-
-        return new UploadedFile($tmp, $size, $error, $original, $type);
-    }
-
-    /**
-     * Diverse the $_FILES into a consistent result.
-     *
-     * @param  array $uploaded
-     * @return array
-     */
-    protected static function diverse(array $uploaded)
-    {
-        $result = array();
-
-        foreach ($uploaded as $file => $item) {
-            foreach ($item as $key => $value) {
-                $diversed = (array) $value;
-
-                $result[$file][$key] = $diversed;
-            }
-        }
-
-        return $result;
+        return $normalized;
     }
 }